Hide "Advanced Settings" settings item [BB-9081] (#1252)
* refactor: convert header utils to hooks

* feat: hide advanced settings button if a user doesn't have access

(cherry picked from commit 45c68d6ca40cf2250f6b8b10fa58afb435561114)
