+++ src/CourseAuthoringPage.jsx
@@ -11,6 +11,7 @@ import { fetchCourseDetail } from './data/thunks';
 import { useModel } from './generic/model-store';
 import NotFoundAlert from './generic/NotFoundAlert';
 import PermissionDeniedAlert from './generic/PermissionDeniedAlert';
+import { fetchStudioHomeData } from './studio-home/data/thunks';
 import { getCourseAppsApiStatus } from './pages-and-resources/data/selectors';
 import { RequestStatus } from './data/constants';
 import Loading from './generic/Loading';
@@ -44,6 +45,10 @@ const CourseAuthoringPage = ({ courseId, children }) => {
   useEffect(() => {
     dispatch(fetchCourseDetail(courseId));
   }, [courseId]);
+
+  useEffect(() => {
+    dispatch(fetchStudioHomeData());
+  }, []);
 
   const courseDetail = useModel('courseDetails', courseId);
